# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/organizations/{id}/permissions": {
      "get": {
        "operationId": "PermissionsController_read[0]",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso; sessão plena e permissão atual na organização. Namespace admin restrito ao Master.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "catalog": {
                      "type": "object",
                      "properties": {
                        "organization.read": {
                          "type": "string",
                          "enum": [
                            "Visualizar organização"
                          ]
                        },
                        "organization.update": {
                          "type": "string",
                          "enum": [
                            "Editar e ativar/desativar organização"
                          ]
                        },
                        "members.read": {
                          "type": "string",
                          "enum": [
                            "Consultar usuários"
                          ]
                        },
                        "members.create": {
                          "type": "string",
                          "enum": [
                            "Cadastrar usuários"
                          ]
                        },
                        "members.link": {
                          "type": "string",
                          "enum": [
                            "Vincular contas existentes"
                          ]
                        },
                        "members.roles": {
                          "type": "string",
                          "enum": [
                            "Alterar cargos"
                          ]
                        },
                        "members.remove": {
                          "type": "string",
                          "enum": [
                            "Remover vínculos"
                          ]
                        },
                        "permissions.manage": {
                          "type": "string",
                          "enum": [
                            "Administrar permissões"
                          ]
                        }
                      },
                      "required": [
                        "organization.read",
                        "organization.update",
                        "members.read",
                        "members.create",
                        "members.link",
                        "members.roles",
                        "members.remove",
                        "permissions.manage"
                      ],
                      "additionalProperties": false
                    },
                    "roles": {
                      "type": "object",
                      "properties": {
                        "ORGANIZATION_ADMIN": {
                          "type": "object",
                          "properties": {
                            "protected": {
                              "type": "boolean"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "enum": [
                                  "organization.read",
                                  "organization.update",
                                  "members.read",
                                  "members.create",
                                  "members.link",
                                  "members.roles",
                                  "members.remove",
                                  "permissions.manage"
                                ]
                              }
                            }
                          },
                          "required": [
                            "protected",
                            "permissions"
                          ],
                          "additionalProperties": false
                        },
                        "MEMBER": {
                          "type": "object",
                          "properties": {
                            "protected": {
                              "type": "boolean"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "enum": [
                                  "organization.read",
                                  "organization.update",
                                  "members.read",
                                  "members.create",
                                  "members.link",
                                  "members.roles",
                                  "members.remove",
                                  "permissions.manage"
                                ]
                              }
                            }
                          },
                          "required": [
                            "protected",
                            "permissions"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "ORGANIZATION_ADMIN",
                        "MEMBER"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "catalog",
                    "roles"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sessão inválida ou sem acesso à operação.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "SESSION_REQUIRED",
                        "PASSWORD_REUSED",
                        "MFA_INVALID",
                        "REAUTHENTICATION_FAILED",
                        "ATTEMPTS_BLOCKED"
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Origem inválida ou sem permissão na organização",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Organização não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflito de vínculo/identidade ou proteção do último administrador",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "413": {
            "description": "Payload maior que 8 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Limite por IP",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Serviço indisponível",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "login_session": []
          }
        ],
        "tags": [
          "Permissions"
        ]
      },
      "post": {
        "operationId": "PermissionsController_save[0]",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "FRONTEND_ORIGIN exata; application/json obrigatório"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "role",
                  "permissions"
                ],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": [
                      "MEMBER"
                    ]
                  },
                  "permissions": {
                    "type": "array",
                    "uniqueItems": true,
                    "items": {
                      "type": "string",
                      "enum": [
                        "organization.read",
                        "organization.update",
                        "members.read",
                        "members.create",
                        "members.link",
                        "members.roles",
                        "members.remove",
                        "permissions.manage"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso; sessão plena e permissão atual na organização. Namespace admin restrito ao Master.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "catalog": {
                      "type": "object",
                      "properties": {
                        "organization.read": {
                          "type": "string",
                          "enum": [
                            "Visualizar organização"
                          ]
                        },
                        "organization.update": {
                          "type": "string",
                          "enum": [
                            "Editar e ativar/desativar organização"
                          ]
                        },
                        "members.read": {
                          "type": "string",
                          "enum": [
                            "Consultar usuários"
                          ]
                        },
                        "members.create": {
                          "type": "string",
                          "enum": [
                            "Cadastrar usuários"
                          ]
                        },
                        "members.link": {
                          "type": "string",
                          "enum": [
                            "Vincular contas existentes"
                          ]
                        },
                        "members.roles": {
                          "type": "string",
                          "enum": [
                            "Alterar cargos"
                          ]
                        },
                        "members.remove": {
                          "type": "string",
                          "enum": [
                            "Remover vínculos"
                          ]
                        },
                        "permissions.manage": {
                          "type": "string",
                          "enum": [
                            "Administrar permissões"
                          ]
                        }
                      },
                      "required": [
                        "organization.read",
                        "organization.update",
                        "members.read",
                        "members.create",
                        "members.link",
                        "members.roles",
                        "members.remove",
                        "permissions.manage"
                      ],
                      "additionalProperties": false
                    },
                    "roles": {
                      "type": "object",
                      "properties": {
                        "ORGANIZATION_ADMIN": {
                          "type": "object",
                          "properties": {
                            "protected": {
                              "type": "boolean"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "enum": [
                                  "organization.read",
                                  "organization.update",
                                  "members.read",
                                  "members.create",
                                  "members.link",
                                  "members.roles",
                                  "members.remove",
                                  "permissions.manage"
                                ]
                              }
                            }
                          },
                          "required": [
                            "protected",
                            "permissions"
                          ],
                          "additionalProperties": false
                        },
                        "MEMBER": {
                          "type": "object",
                          "properties": {
                            "protected": {
                              "type": "boolean"
                            },
                            "permissions": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "enum": [
                                  "organization.read",
                                  "organization.update",
                                  "members.read",
                                  "members.create",
                                  "members.link",
                                  "members.roles",
                                  "members.remove",
                                  "permissions.manage"
                                ]
                              }
                            }
                          },
                          "required": [
                            "protected",
                            "permissions"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "ORGANIZATION_ADMIN",
                        "MEMBER"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "catalog",
                    "roles"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Sessão inválida ou sem acesso à operação.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string",
                      "enum": [
                        "SESSION_REQUIRED",
                        "PASSWORD_REUSED",
                        "MFA_INVALID",
                        "REAUTHENTICATION_FAILED",
                        "ATTEMPTS_BLOCKED"
                      ]
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Origem inválida ou sem permissão na organização",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Organização não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflito de vínculo/identidade ou proteção do último administrador",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "413": {
            "description": "Payload maior que 8 KiB",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Limite por IP",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Serviço indisponível",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "security": [
          {
            "login_session": []
          }
        ],
        "tags": [
          "Permissions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Application Foundation API",
    "description": "Origin exata e application/json obrigatórios em POST. Cookie opaco HttpOnly. Ver README para estados, rate limits e respostas.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "login_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "login_session"
      }
    },
...
  }
}